+++ thunder-core/src/main/java/network/thunder/core/communication/nio/handler/low/MessageObjectToByteHandler.java
@@ -1,4 +1,4 @@
-package network.thunder.core.communication.nio;
+package network.thunder.core.communication.nio.handler.low;
 
 import com.google.gson.Gson;
 import io.netty.buffer.ByteBuf;
